refactor: use float64 for cpu and ram usage percentage

+++ internal/schema/stats.go
@@ -2,6 +2,6 @@ package schema
 
 type Stats struct {
 	ID  string
-	CPU float32
-	RAM uint64
+	CPU float64
+	RAM float64
 }
